fix: resolve relation beetween schedule and employee
- fix types for postgresql
- create new entity

+++ Citas.Domain/Entities/Reservation.cs
@@ -10,15 +10,15 @@ namespace Citas.Domain.Entities
         public required Employee Employee { get; set; }
         public required DateTimeRange RangeTime { get; set; }
 
-       public float Price { get; set; }
+       public double Price { get; set; }
 
        public string? ClientComment { get; set; }
 
-       public int? RatingFromClient { get; set; }
+       public float? RatingFromClient { get; set; }
     
        public string? EmployeeComment { get; set; }
 
-       public int? RatingFromEmployee { get; set; }
+       public float? RatingFromEmployee { get; set; }
 
        public required EReservationState State { get; set; }
     }
